+++ cli/src/commands/parse/mod.rs
@@ -84,11 +84,16 @@ impl Statistics {
     }
 }
 
-pub fn get_files_in_directory(directory: &PathBuf, extension: &str) -> Result<Vec<DirEntry>> {
+pub fn get_files_in_directory(
+    directory: &PathBuf,
+    target_extension: &str,
+) -> Result<Vec<DirEntry>> {
     Ok(std::fs::read_dir(directory)?
         .filter_map(|path| {
             let path = path.ok()?;
-            if path.path().extension().is_some_and(|msg| msg == extension) {
+            if path.path().extension().is_some_and(|extension| {
+                *extension.to_ascii_lowercase() == *target_extension.to_ascii_lowercase()
+            }) {
                 Some(path)
             } else {
                 None