+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+- Fixes issue when getting streams that have multiple filters on single user property
+
 # v0.24.0
 - BREAKING: the `--context` option is now required. Users need to opt
   out if they don't want to provide this for every command