+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# v0.25.0
 - Fixes issue when getting streams that have multiple filters on single user property
 - Fixes issue where upper case file names would not be matched in `parse` 
 - Reduce batch size when deleting comment batches